Adding dockerfile and maven wrapper.

+++ src/main/kotlin/com/verifi/ramon/weather/domain/Wind.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class Wind (
 
 	@SerializedName("speed") val speed : Double,
-	@SerializedName("deg") val deg : Int
+	@SerializedName("deg") val deg : Double
 )
